requirements.txt
torch>=2.0.0
pytorch-lightning>=2.1.0
litdata>=0.1.0
einops>=0.6.0
wandb
numpy
pandas
tqdm
matplotlib
plotly
scikit-learn
typing-extensions
pyyaml
huggingface-hub
transformers
datasets
fair-esm
biopython
dm-tree
hydra-core
omegaconf